pyproject.toml
[project]
name = "aidialogs"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

